+++ packages/calendar/src/components/days-table/Component.tsx
@@ -1,8 +1,7 @@
 /* eslint-disable complexity */
-import React, { FC, RefCallback, useCallback, useEffect, useRef } from 'react';
+import React, { FC, RefCallback, useCallback, useRef } from 'react';
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
 import cn from 'classnames';
-import { getTime } from 'date-fns';
 import isEqual from 'date-fns/isEqual';
 import isLastDayOfMonth from 'date-fns/isLastDayOfMonth';
 import isSameDay from 'date-fns/isSameDay';
@@ -81,16 +80,6 @@ export type DaysTableProps = {
      * Рендерит компонент, обернутый в Transition
      */
     withTransition?: boolean;
-
-    /**
-     * При клике на месяц будут выбраны все доступные дни месяца
-     */
-    monthClicked?: boolean;
-
-    /**
-     * Обработчик выбора даты
-     */
-    onChange?: (date?: number, dateTo?: number) => void;
 };
 
 export const DaysTable: FC<DaysTableProps> = ({
@@ -105,13 +94,9 @@ export const DaysTable: FC<DaysTableProps> = ({
     responsive,
     shape = 'rounded',
     withTransition = true,
-    monthClicked = false,
-    onChange,
 }) => {
     const activeMonthRef = useRef(activeMonth);
     const directionRef = useRef<'right' | 'left' | undefined>();
-    const currentStartMonth = useRef<number>();
-    const currentEndPeriod = useRef<number>();
 
     activeMonthRef.current = activeMonth;
 
@@ -176,16 +161,6 @@ export const DaysTable: FC<DaysTableProps> = ({
         const handleDayClick = (e: React.MouseEvent<HTMLTableDataCellElement>) => {
             if (!day.disabled) onClick(e);
         };
-
-        if (dayIdx === 0) {
-            currentStartMonth.current = startOfMonth(day.date).getTime();
-        }
-
-        if (isToday(day.date)) {
-            currentEndPeriod.current = getTime(day.date);
-        } else if (!currentEndPeriod.current && isLastDayOfMonth(day.date)) {
-            currentEndPeriod.current = getTime(day.date);
-        }
 
         return (
             // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-noninteractive-element-interactions
@@ -252,13 +227,6 @@ export const DaysTable: FC<DaysTableProps> = ({
     );
 
     const renderMonth = () => <tbody>{weeks.map(renderWeek)}</tbody>;
-
-    useEffect(() => {
-        if (monthClicked && onChange) {
-            onChange(currentStartMonth.current, currentEndPeriod.current);
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [monthClicked]);
 
     return (
         <table