+++ packages/custom-picker-button/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @alfalab/core-components-custom-picker-button
 
+## 2.3.6
+
+### Patch Changes
+
+<sup><time>04.09.2024</time></sup>
+
+### [#1353](https://github.com/core-ds/core-components/pull/1353)
+
+-   Обновлены наименования переменных отступов
+
+-   Обновлены зависимости
+    -   select@17.10.1
+    -   picker-button@11.7.18
+    -   custom-button@3.6.5
+
 ## 2.3.5
 
 ### Patch Changes