+++ packages/input/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 15.0.5
+
+### Patch Changes
+
+<sup><time>04.09.2024</time></sup>
+
+### [#1324](https://github.com/core-ds/core-components/pull/1324)
+
+-   Кнопка "Очистить" оборажается в случае `clear={true}` и `disableUserInput={true}`
+
+<sup><time>04.09.2024</time></sup>
+
+### [#1354](https://github.com/core-ds/core-components/pull/1354)
+
+-   Обновлены наименования переменных отступов
+
+-   Обновлены зависимости
+    -   button@11.5.5
+    -   form-control@12.2.8
+
 ## 15.0.4
 
 ### Patch Changes