+++ packages/list-header/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 3.2.8
+
+### Patch Changes
+
+<sup><time>04.09.2024</time></sup>
+
+### [#1354](https://github.com/core-ds/core-components/pull/1354)
+
+-   Обновлены наименования переменных отступов
+
+-   Обновлены зависимости
+    -   typography@4.9.0
+
 ## 3.2.7
 
 ### Patch Changes