+++ packages/tag/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 8.3.1
+
+### Patch Changes
+
+<sup><time>04.09.2024</time></sup>
+
+### [#1356](https://github.com/core-ds/core-components/pull/1356)
+
+-   Обновлены наименования переменных отступов
+
 ## 8.3.0
 
 ### Minor Changes